[banjo] Make parameters inout if they are tagged with "Mutable".
It turns out that in the async case, some output parameters are passed to
the method formatting input parameters. This means that output parameters
should be handled the same way as input parameters. However, marking an
output parameter as "InOut" does not make much sense, which is why we allow
them to be marked as "Mutable" instead.

Bug: 67708
Change-Id: I717a042de3b9543ad738da724205300494bb83ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474820
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: P.Y. Laligand <pylaligand@google.com>