[Bind] Encode the symbol table for new bytecode format
Add headers and the symbol table to the new bytecode
encoding. This CL modifies the Condition enum so that it
stores Symbol values instead of u32.

For additional clean up, this CL deletes c_generation.rs
instead of updating the Condition code in the file since
the file is unused.

Bug: 67440
Test: Added unit tests in encode_bind_program_v2
Change-Id: I5cd3a65aa53641735b9fb3105edb8447cc4c8252
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472842
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>