[bind] Encode the abort instructions
Add bytecode encoding for abort instructions. The Encoder
struct now contains an iterator for the instructions and will
iterate through them when encode_inst_block() is called.

The test verify_next_u32() function is also updated so that instead
of receiving a byte array for u32 values, it'll take in a u32 value
and convert it into bytes.

Bug: 67440
Tests: Added unit tests for the abort instructions and values
Change-Id: I0975689b2871160c381410fc9cdcf471692bb6d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474627
Commit-Queue: Sarah Chan <spqchan@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>