[bind] Refactor tests in encode_bind_program_v2
Refactored the tests in encode_bind_program_v2 so
that it's easier to read and write them, especially
when the tests become more complex in the future.

This CL introduce constants for the number of bytes
in instructions so that the tests can use them to
calculated the expected results.

It also introduces helper methods to verify the
instructions and the headers.

Bug: 67440
Change-Id: Ia1259c55856523a304b9b939ad07351e55c66e91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478638
Reviewed-by: Alex Legg <alexlegg@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>