[bind] Encode jump statements
Encode jump statements in the new bytecode format.
Since the jump statements require calculating the
jump offsets, a map that contains the label
ID, the label index and the jump instructions
is added to the Encoder struct.

The map is updated when the bind program encounters
a jump instruction or label. Once the Encoder
finished encoding all of the instructions, it iterates
through the map and update all of the encoded jump
instructions with the calculated jump offsets.

Bug: 67440
Tests: Added unit tests
Change-Id: Ibe5c566ece4ada5a8f79d45eaf2ebd487343ba33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/479117
Commit-Queue: Sarah Chan <spqchan@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>