[pci] Enable bus mastering if an MSI mode is enabled
Message signaled interrupts require bus mastering capabilities on the
device to generate messages. This was left up to the device drivers, but
is easy to overlook and hard to debug without PCI knowledge. This change
enables bus mastering when MSI is requested in both the kernel and
userspace PCI drivers. Some locking and annotations were changed to
allow this path, and only the userspace driver has a test framework so
no testing has been done for the kernel driver besides manual
verification.

Bug: 68830
Change-Id: Ia3e2df0bee8f460402ca02aeea53c7bf4c9e5176
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482462
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>