[banjo][c] Separate helper declarations from other declarations.
The ordering of protocol declarations is looser in FIDL, meaning a
protocol A may be declared before a protocol B it depends on.
Since helper methods require full types to be available, move these
methods to the very end of the generated header so that all necessary
declarations are available.

Bug: 68157
Change-Id: I520a2f26cae5d2cc5a4b49cfb5f0f3e7a9b2bc07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483381
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: P.Y. Laligand <pylaligand@google.com>