[zircon-crypto] Start moving to //src/security/fcrypto
This will require a three-step transition to avoid breaking source
compat for an internal repo:

* Add `//src/security/fcrypto` target/code and deprecate
  `//zircon/system/ulib/zircon-crypto`
* Migrate internal, non-fuchsia.git usages
* Remove `//zircon/system/ulib/zircon-crypto` and transitional
  target/headers

This patch is step 1.

We shift the name from `zircon-crypto` to `fcrypto` because this
library, despite its name, is not used in the zircon kernel at all.
Its users are limited to zxcrypt (and other programs that link it)
and an internal tool.

We avoid the name `crypto` because that might conflict in library name
with boringssl's `libcrypto`, hence `fcrypto` for "Fuchsia crypto".

This commit moves tests, fuzzers, and all in-tree usages of
//zircon/system/ulib/zircon-crypto, but leaves a transitional group()
target at //zircon/system/ulib/zircon-crypto which redirects to the new
source_set along with including a transitional config for include_dirs
with include stubs to preserve #include compability for purposes of soft
transitions across repos as described above.

Change-Id: If7651e49f8f6b0dde8433abd93be31cd07579dd5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483441
Commit-Queue: Drew Fisher <zarvox@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>