[tee][optee] Apply deadline profile
Some trusted application workloads have certain performance and latency
requirements to be completed. Currently, the OP-TEE driver is unaware of
the workload requirements of individual tasks, so the changes in this CL
are temporary until we can give clients some control over describing the
workload to the TEE driver.

Bug: 67050
Change-Id: I8aa4ace6e7386bf292230a017d76247f90d8cd63
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486581
Reviewed-by: Jonathan Bailey <jonathanbailey@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: RJ Ascani <rjascani@google.com>