[devices] Ensure the root node is removed last in Node::Remove()
When removing a sub-tree, we may already be in the context of an unbind
operation. This means that we need to carefully handle the removal of
the root node in the sub-tree, so that all its children are removed
first.

Bug: 33183
Change-Id: Id0c1763973771d503747d52473fafa857087acb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488097
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>