[rust] Set output_name for rustc_binary calls experiencing normalization
In preparation for a change to rustc_binary which will make output_name
default to target name without "-" -> "_" normalization, set output_name
on every binary currently experiencing that.

This change is a no-op; the generated binary names are staying the same;
the current binary files have _ in the name.

This CL is not meant to imply that the binaries _should_ have `_` or `-`
in the name, it's merely a reflection of the current reality. If you
want the binary to have "-" in the name, you can actually do that now by
setting output_name as of http://fxrev.dev/450757 , but be aware that
anywhere that references the name of the binary (.cmx and .cml files,
for example) will need to be updated to reflect the new name.

Bug: 64100
Change-Id: I1b1a359ee53e4ec80c424d399f247ff8e6b845a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486477
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: Dan Johnson <computerdruid@google.com>