[devices] Wait on peer closed of the Driver protocol
Closing of the channel associated with the Driver protocol is used to
indicate to the Driver Runner that the driver has stopped executing
within a Driver Host. We should watch for that in the Driver Runner and
perform the appropriate clean up.

Bug: 33183
Change-Id: Ia38af19aa54ccfcefb9d0ab8b2316eb89a4a59d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489457
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>