[bind] Define fuchsia.platform DID for XHCI
This adds the corresponding value to PDEV_DID_USB_XHCI_COMPOSITE.

Right now, we have two DIDs for the xHCI controllers: a DID_USB_XHCI and
DID_USB_XHCI_COMPOSITE.  Of the two, the composite variant is
overwhelmingly used.  We don't actually need to DIDs here, and to clean
this up I'm assuming the more commonly used value, and leaving the other
(which is only currently relevant to the vim2) to be cleaned up later.

Change-Id: I6966c301fcdb631ca13bfdc3e7852f95d5e87874
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489277
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Sam Hansen <hansens@google.com>