[tee][optee] Instrument yielding SMC calls
The OP-TEE driver makes a yielding SMC call whenever it is attempting to
communicate with a Trusted Application. Yielding SMC calls in the secure
world will operate with interrupts enabled and will return to the
non-secure world when one occurs. In order to detect whether this return
is taking too long, this CL adds a warning log whenever the peak SMC
call duration for a Trusted App call exceeds a given threshold. This
threshold is currently set to 10ms. Overall, the Trusted App call could
be much longer, but we are only trying to detect if a single SMC call
exceeds the duration.

Bug: 69037
Change-Id: I87b23dd5535ce01be0be2890dd4a214967b09637
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489889
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Dmitry Yatsushkevich <dmitryya@google.com>