[fidl][llcpp] UnownedClientEnd can return unowned_channel
Initially when defining UnownedClientEnd, we only used zx_handle_t
because zx::unowned_channel had a lot of baggage for what is a simple
handle number.

It turns out that there are lots of existing code that expect
zx::unowned_channel. It seems worthwhile to support returning a
zx::unowned_channel from an UnownedClientEnd. The getter that only
returns a zx_handle_t was renamed to `handle()`.

User code that obtains a handle just to call fdio_clone_service on it
are migrated to use `service::Clone`. In other cases, they are migrated
to use the `handle()` getter.

Change-Id: I74f3f66b5515c055e36d066c337e7a9738927eef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489447
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Vincent Belliard <vbelliard@google.com>