[dwc][usb] Flush all TX endpoints when handling a RESET
The following bugs seem to be symptoms of an improperly handled reset:
 + fxb/66889
 + fxb/60670
 + fxb/57394

The DWC2 TRM suggests that flushing all FIFOs is one of the steps
that should be taken when handling a RESET however our driver didn't
appear to take this step.

Tested on Astro, Sherlock and Vim3.

Change-Id: Ie914cc06d744d53705939e6476b0f85f2b91c6a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493398
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>