[virtio] Migrate to zx_system_get_page_size
Migrates usages of page size constants to zx_system_get_page_size

The virtio rings are not required to be sized at precisely the system
page size, the only requirement is that they are contiguous in the
physical view of memory. As such they could be sized smaller. However,
since the existing implementation chose to tie the ring size to the
system page size and not make it smaller, this preserves that choice.

Bug: 67899

Change-Id: I171b40b78dc75904fb5888f51313e1931daf3d90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497998
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>