[rtc] Cease maintaining the Kernel UTC clock.
As per RFC-0008 we're in the process of removing the original
`zx_clock_get` and `zx_clock_adjust` syscalls along with the kernel UTC
clock.

kcounters show that we've successfully removed the last usages of the
kernel UTC clock and so this CL stops setting the kernel UTC clock. The
"fallback" rtc driver is now no longer doing anything useful and will be
removed in a future CL.

Bug: 31359
Change-Id: I3c01c5cdc3e5e4987a727fa9a3f6bee3054b6ade
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495159
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>