[aml][cpu] CPU Device now correctly gets CPU count from metadata.
CPU core count was previously reported through zx_system_get_num_cpus
which returned the total number of logical cores in the system.

This is only correct on systems with a single performance domain.

This patch facilitates the ability to report number of cores on a per
performance domain basis.

Test: aml-cpu-test
Change-Id: I0a52a0390766d455e0f585ec911e66ea7aab8745
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500019
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>
Reviewed-by: Patrick Shickel <pshickel@google.com>
Reviewed-by: Onath Claridge <claridge@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>