[devices] Move drivers and driver_hosts collections to bootstrap
This allows us begin splitting the drivers collection, based on the
dependencies of the collections, and thus break dependency cycles.

Specifically, we can begin to provide different resolvers to different
driver collections, so that there is a strong dependency between the
drivers and the file systems that they are read from.

Bug: 33183
Change-Id: I7d15268378a42dcb4d1da8da6ceaeb90e7ba9314
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505198
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>