[dev-operation] Migrate off fbl.
We need dev-operation for writing out-of-tree drivers that use usb,
such as the example ftdi driver.

We do not want to include fbl in the SDK, but there is no
std equivalent for the intrusive_double_list. The options proposed
were to downgrade the ftdi driver to use c-style usb requests,
or make a branch of the required fbl libraries. Branching
fbl seems to be the easiest path forward, and since we eventually
expect to deprecate dev-operation, this should not increase tech debt.

I have copied the required fbl headers into lib/operation/helpers,
and changed the namespace to operation. I also changed operation.h's
use of fbl locks to the std locks.

Bug: 71471
Change-Id: I148c6bf71c13bce10d99533cda16ce672d97e368
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506687
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>