[inotify] Add localfileystem,remote-io, zxio support.
This CL makes changes to remote/zxio layers as well
as local-filesystem layers to make the FIDL call to
VFS via DirectoryIo protocol.

Also changed the ordering of the arguments to be consistent
throughout the various calls to AddInotifyFilter.

Also, we change zx::socket to zx_handle_t to pass via
the zxio layer which needs to be compatible with the
C interface and then later wrap it up back into
zx::socket for the fidl calls.

The readv method is intentionally left unimplemented,
and will be done as a followup CL with tests.

Testing: Added fake server to emulate DirectoryIO protocol
         to test if we are calling the FIDL correctly from the
         namespace.

Change-Id: Ieeefe314182a99f0c8f76d98f58e0b5625ca06a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488778
Commit-Queue: Manali Bhutiyani <manalib@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
API-Review: Adam Barth <abarth@google.com>