[banjo] Add heuristic for generating callbacks
FIDL IR does not contain attributes for external types, so we cannot
know if the `ddk-callback` value is set for the `BanjoProtocol`
attribute. As a result, we can't codegen the appropriate struct name
with the fidlgen backend. As a workaround, we've added a heuristic
whereby the Callback suffix in the protocol type name also signals that
it is a `ddk-callback`.

Tested: fx test src/devices/tools/
Change-Id: Id335d7eb69d1be296418dfcc56475083d160c860
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/512678
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>