[nelson][aml-spi] Use deadline profile for Amlogic SPI driver
The next change to the SPI driver will enable interrupt-driven IO. In
order to use interrupts and still meet the timing requirements, a
deadline profile has to be applied to the thread handling the SPI
transfers. This change adds metadata to specify the profile parameters,
and applies the profile in the first transfer request.

Bug: 74692
Test: Let a Nelson running the product config idle for 45 minutes, saw
      no radar driver timeouts. Without this change there were 27
      timeouts in the same interval.
Test: ot-radio driver initialized on Sherlock.
Change-Id: I64fdcb8055300332f8e4a87847fdde933ade58a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519998
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>