[pci] Use DDKTL for kpci
For some of the fragment work that needs to be done it will be
far easier if Kernel PCI and User PCI have a similar base. This
change brings forward the legacy C kpci driver over to DDKTL/C++,
but tries to make as few functional changes as possible otherwise.

Bug: 32978
Change-Id: Ib3ee9037237a5828d82f6b2ed70894b6bf12e40f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516720
Commit-Queue: Christopher Anderson <cja@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>