[bindc] Add fuchsia.BIND_PCI_COMPONENT
To migrate the PCI Bus Driver from simple devices to composite devices
made up of multiple fragments we need a way to designate whether a
device driver has been migrated, as well as whether a device is for a
fragment or a simple device driver. This change allows us to mark a
driver as a simple or composite driver, as well as makes it possible to
denote a device as a fragment used only for building a composite out of
while other drivers still want to bind to it.

For this CL all PCI device drivers are "NOT_MIGRATED", aka simple.

Bug: 74834
Change-Id: Ia8f3918f526b8d10cba8f526eacd9eeb969e3d9b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519042
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Christopher Anderson <cja@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>