[devices] Wait for drivers to stop when removing nodes
When removing nodes from the driver topology, if a node is bound to a
driver, wait for the driver to stop before removing the node.

This ensures that drivers can not be stopped out of order when run in
different driver hosts.

Bug: 74571
Bug: 33183
Change-Id: I6511f1fde45ff4e73a8a56f79817ae176131d69d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524684
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>