[radar][bin] Fix help handling and choose a better run time default
Passing -h should only show the help message and not try to read radar
bursts. Return an error code when -h is passed to exit early.

Also, choose 1 second as the default amount of time to read bursts.
Running for 10 seconds (the previous default) with no output could be
confusing for first-time users of the tool.

Bug: 67567
Test: radarutil -h showed the help message and didn't read bursts.
Test: fx test radarutil-test
Change-Id: I92845b9870490d4cc7e80d69cf1e1e1bc1c7eaee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525328
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>