[devices] Reduce the size of a Node
Reduce the size of the initial capacity of the FIDL allocator embedded
inside a Node so that the overall structure drops in size from 784 bytes
to 400 bytes. This means that 512 Nodes will take up roughly 200kb,
instead of 400kb. If the system allocator rounds up to the nearest power
of 2, the savings will be proportional.

Bug: 33183
Change-Id: Ia68b76bdfd48931d0cdd1e50dba25b9879bac31d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526549
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>